Refactoring app files into packages
