+++ app/src/main/java/com/sample/spaceridegalley/imagedetails/domain/ImageDetailUseCase.kt
@@ -1,0 +1,7 @@
+package com.sample.spaceridegalley.imagedetails.domain
+
+import com.sample.spaceridegalley.imagedetails.domain.model.DetailItemModel
+
+interface ImageDetailUseCase {
+    suspend fun getImageItemList() : List<DetailItemModel>
+}
