--- app/src/main/java/com/sample/spaceridegalley/imagedetails/view/ImageDetailUseCase.kt
@@ -1,5 +1,0 @@
-package com.sample.spaceridegalley.imagedetails.view
-
-interface ImageDetailUseCase {
-    suspend fun getImageItemList() : List<DetailItemModel>
-}
