+++ app/src/main/java/com/sample/spaceridegalley/imagedetails/view/ImageDetailsViewPagerAdapter.kt
@@ -1,20 +1,17 @@
 package com.sample.spaceridegalley.imagedetails.view
 
-import android.annotation.SuppressLint
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
-import androidx.navigation.findNavController
 import androidx.recyclerview.widget.RecyclerView
 import androidx.swiperefreshlayout.widget.CircularProgressDrawable
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.engine.DiskCacheStrategy
-import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions
 import com.bumptech.glide.request.RequestOptions
 import com.sample.spaceridegalley.R
 import com.sample.spaceridegalley.databinding.FragmentImageDetailSingleBinding
-import com.sample.spaceridegalley.databinding.FragmentImageListSingleBinding
+import com.sample.spaceridegalley.imagedetails.domain.model.DetailItemModel
 
 
 class ImageDetailsViewPagerAdapter(private val values: MutableList<DetailItemModel>) : RecyclerView.Adapter<ImageDetailsViewPagerAdapter.ViewHolder>() {
