--- app/src/main/java/com/sample/spaceridegalley/imagelist/ImageListUseCase.kt
@@ -1,5 +1,0 @@
-package com.sample.spaceridegalley.imagelist
-
-interface ImageListUseCase {
-    suspend fun getImageItemList() : List<ListItemModel>
-}
