+++ app/src/main/java/com/sample/spaceridegalley/imagelist/domain/ImageListUseCase.kt
@@ -1,0 +1,7 @@
+package com.sample.spaceridegalley.imagelist.domain
+
+import com.sample.spaceridegalley.imagelist.domain.model.ListItemModel
+
+interface ImageListUseCase {
+    suspend fun getImageItemList() : List<ListItemModel>
+}
