+++ app/src/main/java/com/sample/spaceridegalley/imagelist/domain/model/ListItemModel.kt
@@ -1,3 +1,3 @@
-package com.sample.spaceridegalley.imagelist
+package com.sample.spaceridegalley.imagelist.domain.model
 
 data class ListItemModel(val lowQUrl : String, val title : String)
