+++ app/src/main/java/com/sample/spaceridegalley/imagelist/view/ImageListViewModel.kt
@@ -1,15 +1,14 @@
 package com.sample.spaceridegalley.imagelist
 
-import android.app.appsearch.SearchResult
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.sample.spaceridegalley.imagelist.domain.ImageListUseCase
+import com.sample.spaceridegalley.imagelist.domain.model.ListItemModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
