+++ app/src/main/res/layout/fragment_image_detail.xml
@@ -16,7 +16,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        tools:context=".imagelist.ImageListFragment"
+        tools:context=".imagelist.view.ImageListFragment"
         tools:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
         tools:listitem="@layout/fragment_image_detail_single"
         tools:spanCount="2" />
